requirements.txt
kafka-python
pyspark
pandas
numpy
scikit-learn
setuptools
matplotlib
streamlit